+++ src/Exceptions/ExceptionsServiceProvider.php
@@ -3,13 +3,14 @@
 namespace Engelsystem\Exceptions;
 
 use Engelsystem\Container\ServiceProvider;
-use Engelsystem\Exceptions\Handler as ExceptionHandler;
 
 class ExceptionsServiceProvider extends ServiceProvider
 {
     public function register()
     {
-        $errorHandler = $this->app->make(ExceptionHandler::class);
+        $errorHandler = $this->app->make(BasicHandler::class);
+        $errorHandler->register();
         $this->app->instance('error.handler', $errorHandler);
+        $this->app->bind(Handler::class, 'error.handler');
     }
 }
