+++ tests/Unit/Exceptions/ExceptionsServiceProviderTest.php
@@ -2,8 +2,9 @@
 
 namespace Engelsystem\Test\Unit\Exceptions;
 
+use Engelsystem\Exceptions\BasicHandler as ExceptionHandler;
 use Engelsystem\Exceptions\ExceptionsServiceProvider;
-use Engelsystem\Exceptions\Handler as ExceptionHandler;
+use Engelsystem\Exceptions\Handler;
 use Engelsystem\Test\Unit\ServiceProviderTest;
 use PHPUnit_Framework_MockObject_MockObject;
 
@@ -18,10 +19,11 @@ class ExceptionsServiceProviderTest extends ServiceProviderTest
         $exceptionHandler = $this->getMockBuilder(ExceptionHandler::class)
             ->getMock();
 
-        $app = $this->getApp();
+        $app = $this->getApp(['make', 'instance', 'bind']);
 
         $this->setExpects($app, 'make', [ExceptionHandler::class], $exceptionHandler);
         $this->setExpects($app, 'instance', ['error.handler', $exceptionHandler]);
+        $this->setExpects($app, 'bind', [Handler::class, 'error.handler']);
 
         $serviceProvider = new ExceptionsServiceProvider($app);
         $serviceProvider->register();
